refactor(services): move client logic to separate component
